+++ src/app/student/specific/pages/specific-quest-map/specific-quest-map.component.css
@@ -13,6 +13,18 @@ h3:after {
     -moz-box-shadow:inset 0 1px 1px rgba(0, 0, 0, .05);
 }
 
+.filled-circle {
+    padding: 2px 11px;
+    border-radius: 100%;
+    background-color: #f38e23;
+}
+
+.topcorner{
+    position:absolute;
+    top:0;
+    right:0;
+   }
+
 .modal-header h4 {
     margin: 0 auto;
 }
