+++ src/app/student/specific/pages/specific-quest-map/specific-quest-map.component.html
@@ -1,6 +1,11 @@
 <!--Graph for quest map-->
 <h3>Quest Map</h3>
 
+<div class="topcorner"> 
+	haha
+	<span class="filled-circle">hello</span>
+</div>
+
 <div class="row" *ngIf="quests">
 	<div style="display: block">
 		<canvas id="quest-map" baseChart width="650" height="300"></canvas>
