+++ src/app/toast-option.ts
@@ -6,5 +6,6 @@ export class CustomOption extends ToastOptions {
 	animate = 'flyRight'; // you can override any options available
     newestOnTop = false;
 	showCloseButton = true;
-	toastLife = 3000;
+	enableHTML = true;
+	maxShown = 1;
 }
